requirements.txt
openai
pandas
pyodbc
python-pptx
autogen
python-dotenv
fastapi
uvicorn

